# docs/swagger.json
{
    "swagger": "2.0",
    "info": {
        "description": "API user management Server by Fiber | Doc by Swagger.",
        "title": "Talardnad API",
        "contact": {
            "name": "admin",
            "url": "http://subalgo.com/support",
            "email": "admin@subalgo.com"
        },
        "license": {
            "name": "Apache 2.0",
            "url": "http://www.apache.org/licenses/LICENSE-2.0.html"
        },
        "version": "1.0"
    },
    "host": "localhost:3000",
    "basePath": "/api/v1",
    "paths": {
        "/auth/login": {
            "post": {
                "description": "Login with the provided credentials",
                "consumes": [
                    "application/json"
                ],
                "produces": [
                    "application/json"
                ],
                "tags": [
                    "auth"
                ],
                "summary": "Login",
                "parameters": [
                    {
                        "description": "Login data",
                        "name": "login",
                        "in": "body",
                        "required": true,
                        "schema": {
                            "$ref": "#/definitions/entities.LoginRequest"
                        }
                    }
                ],
                "responses": {
                    "200": {
                        "description": "OK",
                        "schema": {
                            "$ref": "#/definitions/entities.LoginResponse"
                        }
                    },
                    "400": {
                        "description": "Invalid input",
                        "schema": {
                            "type": "string"
                        }
                    },
                    "401": {
                        "description": "Invalid credentials",
                        "schema": {
                            "type": "string"
                        }
                    }
                }
            }
        },
        "/auth/provider/login": {
            "post": {
                "description": "Login for market providers with the provided credentials",
                "consumes": [
                    "application/json"
                ],
                "produces": [
                    "application/json"
                ],
                "tags": [
                    "auth"
                ],
                "summary": "Provider Login",
                "parameters": [
                    {
                        "description": "Provider Login data",
                        "name": "login",
                        "in": "body",
                        "required": true,
                        "schema": {
                            "$ref": "#/definitions/dtos.ProviderLoginRequest"
                        }
                    }
                ],
                "responses": {
                    "200": {
                        "description": "OK",
                        "schema": {
                            "$ref": "#/definitions/dtos.ProviderLoginResponse"
                        }
                    },
                    "400": {
                        "description": "Invalid input",
                        "schema": {
                            "type": "string"
                        }
                    },
                    "401": {
                        "description": "Invalid credentials",
                        "schema": {
                            "type": "string"
                        }
                    }
                }
            }
        },
        "/auth/provider/register": {
            "post": {
                "description": "Register a new market provider with the provided data",
                "consumes": [
                    "application/json"
                ],
                "produces": [
                    "application/json"
                ],
                "tags": [
                    "auth"
                ],
                "summary": "Register Provider",
                "parameters": [
                    {
                        "description": "Register provider request",
                        "name": "register",
                        "in": "body",
                        "required": true,
                        "schema": {
                            "$ref": "#/definitions/dtos.MarketProviderRequest"
                        }
                    }
                ],
                "responses": {
                    "200": {
                        "description": "OK",
                        "schema": {
                            "$ref": "#/definitions/entities.MarketProvider"
                        }
                    },
                    "400": {
                        "description": "Failed to register provider",
                        "schema": {
                            "type": "string"
                        }
                    },
                    "409": {
                        "description": "Provider email already exists",
                        "schema": {
                            "type": "string"
                        }
                    }
                }
            }
        },
        "/auth/register": {
            "post": {
                "description": "Register a new user with the provided data",
                "consumes": [
                    "application/json"
                ],
                "produces": [
                    "application/json"
                ],
                "tags": [
                    "auth"
                ],
                "summary": "Register",
                "parameters": [
                    {
                        "description": "Register request",
                        "name": "register",
                        "in": "body",
                        "required": true,
                        "schema": {
                            "$ref": "#/definitions/dtos.RegisterRequest"
                        }
                    }
                ],
                "responses": {
                    "200": {
                        "description": "OK",
                        "schema": {
                            "$ref": "#/definitions/dtos.RegisterResponse"
                        }
                    },
                    "400": {
                        "description": "Failed to register user",
                        "schema": {
                            "type": "string"
                        }
                    },
                    "409": {
                        "description": "Email already exists",
                        "schema": {
                            "type": "string"
                        }
                    }
                }
            }
        },
...
        "/markets/create": {
            "post": {
                "description": "Create a new market",
                "consumes": [
                    "application/json"
                ],
                "produces": [
                    "application/json"
                ],
                "tags": [
                    "Market"
                ],
                "summary": "Create a new market",
                "parameters": [
                    {
                        "description": "Market object that needs to be created",
                        "name": "market",
                        "in": "body",
                        "required": true,
                        "schema": {
                            "$ref": "#/definitions/dtos.MarketRequest"
                        }
                    }
                ],
                "responses": {
                    "201": {
                        "description": "Created",
                        "schema": {
                            "$ref": "#/definitions/dtos.MarketResponse"
                        }
                    }
                }
            }
        },
        "/markets/edit/{id}": {
            "patch": {
                "description": "Edit a market",
                "consumes": [
                    "application/json"
                ],
                "produces": [
                    "application/json"
                ],
                "tags": [
                    "Market"
                ],
                "summary": "Edit a market",
                "parameters": [
                    {
                        "type": "string",
                        "description": "Market ID",
                        "name": "id",
                        "in": "path",
                        "required": true
                    },
                    {
                        "description": "Market object that needs to be updated",
                        "name": "market",
                        "in": "body",
                        "required": true,
                        "schema": {
                            "$ref": "#/definitions/dtos.MarketEditRequest"
                        }
                    }
                ],
                "responses": {
                    "200": {
                        "description": "OK",
                        "schema": {
                            "$ref": "#/definitions/entities.Market"
                        }
                    }
                }
            }
        },
...
        "/markets/get/{id}": {
            "get": {
                "description": "Get a market by ID",
                "consumes": [
                    "application/json"
                ],
                "produces": [
                    "application/json"
                ],
                "tags": [
                    "Market"
                ],
                "summary": "Get a market by ID",
                "parameters": [
                    {
                        "type": "string",
                        "description": "Market ID",
                        "name": "id",
                        "in": "path",
                        "required": true
                    }
                ],
                "responses": {
                    "200": {
                        "description": "OK",
                        "schema": {
                            "$ref": "#/definitions/dtos.MarketResponse"
                        }
                    }
                }
            }
        },
        "/markets/provider/get/{id}": {
            "get": {
                "description": "Get a market by Provider ID",
                "consumes": [
                    "application/json"
                ],
                "produces": [
                    "application/json"
                ],
                "tags": [
                    "Market"
                ],
                "summary": "Get a market by Provider ID",
                "parameters": [
                    {
                        "type": "string",
                        "description": "Provider ID",
                        "name": "id",
                        "in": "path",
                        "required": true
                    }
                ],
                "responses": {
                    "200": {
                        "description": "OK",
                        "schema": {
                            "$ref": "#/definitions/dtos.MarketResponse"
                        }
                    }
                }
            }
        },
...
    },
    "definitions": {
...
        "dtos.MarketEditRequest": {
            "type": "object",
            "required": [
                "address",
                "close_time",
                "name",
                "open_time",
                "provider_id"
            ],
            "properties": {
                "address": {
                    "description": "Required, address of the market",
                    "type": "string"
                },
                "close_time": {
                    "description": "Required, closing time in HH:mm format",
                    "type": "string"
                },
                "description": {
                    "description": "Optional, description of the market",
                    "type": "string"
                },
                "image": {
                    "description": "Optional, URL or path to the market image",
                    "type": "string"
                },
                "latitude": {
                    "description": "Optional, latitude coordinate",
                    "type": "string"
                },
                "layout_image": {
                    "description": "Optional, URL or path to the market layout image",
                    "type": "string"
                },
                "longitude": {
                    "description": "Optional, longitude coordinate",
                    "type": "string"
                },
                "name": {
                    "description": "Required, name of the market",
                    "type": "string"
                },
                "open_time": {
                    "description": "Required, opening time in HH:mm format",
                    "type": "string"
                },
                "phone": {
                    "description": "Optional, phone number of the market",
                    "type": "string"
                },
                "provider_id": {
                    "description": "Required, UUID of the provider",
                    "type": "string"
                }
            }
        },
        "dtos.MarketProviderRequest": {
            "type": "object",
            "required": [
                "email",
                "password",
                "phone",
                "username"
            ],
            "properties": {
                "email": {
                    "description": "Required, email address of the provider",
                    "type": "string"
                },
                "password": {
                    "description": "Required, password of the provider",
                    "type": "string"
                },
                "phone": {
                    "description": "Required, phone number of the provider",
                    "type": "string"
                },
                "username": {
                    "description": "Required, username of the provider",
                    "type": "string"
                }
            }
        },
        "dtos.MarketRequest": {
            "type": "object",
            "required": [
                "address",
                "close_time",
                "name",
                "open_time",
                "provider_id"
            ],
            "properties": {
                "address": {
                    "description": "Required, address of the market",
                    "type": "string"
                },
                "close_time": {
                    "description": "Required, closing time in HH:mm format",
                    "type": "string"
                },
                "description": {
                    "description": "Optional, description of the market",
                    "type": "string"
                },
                "image": {
                    "description": "Optional, URL or path to the market image",
                    "type": "string"
                },
                "latitude": {
                    "description": "Optional, latitude coordinate",
                    "type": "string"
                },
                "longitude": {
                    "description": "Optional, longitude coordinate",
                    "type": "string"
                },
                "name": {
                    "description": "Required, name of the market",
                    "type": "string"
                },
                "open_time": {
                    "description": "Required, opening time in HH:mm format",
                    "type": "string"
                },
                "provider_id": {
                    "description": "Required, UUID of the provider",
                    "type": "string"
                }
            }
        },
        "dtos.MarketResponse": {
            "type": "object",
            "properties": {
                "data": {
                    "description": "the data to be returned",
                    "type": "array",
                    "items": {
                        "$ref": "#/definitions/entities.Market"
                    }
                },
                "message": {
                    "description": "message to accompany the response",
                    "type": "string"
                },
                "status": {
                    "description": "success or error",
                    "type": "string"
                }
            }
        },
        "dtos.ProviderLoginRequest": {
            "type": "object",
            "required": [
                "password",
                "username"
            ],
            "properties": {
                "password": {
                    "type": "string",
                    "minLength": 8
                },
                "username": {
                    "type": "string"
                }
            }
        },
        "dtos.ProviderLoginResponse": {
            "type": "object",
            "properties": {
                "access_token": {
                    "type": "string"
                },
                "provider_id": {
                    "type": "string"
                }
            }
        },
        "dtos.RegisterRequest": {
            "type": "object",
            "required": [
                "email",
                "password",
                "phone_number",
                "username"
            ],
            "properties": {
                "email": {
                    "description": "Required, must be a valid email format",
                    "type": "string"
                },
                "password": {
                    "description": "Required, min 8 characters for password",
                    "type": "string",
                    "minLength": 8
                },
                "phone_number": {
                    "description": "Required, adjust based on the expected format=",
                    "type": "string",
                    "maxLength": 15,
                    "minLength": 10
                },
                "username": {
                    "description": "Required, min 3, max 50 characters",
                    "type": "string",
                    "maxLength": 50,
                    "minLength": 3
                }
            }
        },
        "dtos.RegisterResponse": {
            "type": "object",
            "properties": {
                "email": {
                    "type": "string"
                },
                "id": {
                    "type": "string"
                },
                "phone": {
                    "type": "string"
                },
                "username": {
                    "type": "string"
                }
            }
        },
...
        "entities.Booking": {
            "type": "object",
            "properties": {
                "booking_date": {
                    "description": "Changed from Date to BookingDate",
                    "type": "string"
                },
                "created_at": {
                    "type": "string"
                },
                "expires_at": {
                    "type": "string"
                },
                "id": {
                    "type": "string"
                },
                "market_id": {
                    "type": "string"
                },
                "method": {
                    "$ref": "#/definitions/entities.Method"
                },
                "payment": {
                    "$ref": "#/definitions/entities.Payment"
                },
                "price": {
                    "type": "number"
                },
                "slot": {
                    "$ref": "#/definitions/entities.Slot"
                },
                "slot_id": {
                    "type": "string"
                },
                "status": {
                    "$ref": "#/definitions/entities.BookingStatus"
                },
                "updated_at": {
                    "type": "string"
                },
                "vendor": {
                    "$ref": "#/definitions/entities.Vendor"
                },
                "vendor_id": {
                    "type": "string"
                }
            }
        },
...
        "entities.Category": {
            "type": "string",
            "enum": [
                "clothes",
                "food",
                "crafts",
                "produce",
                "electronics",
                "services",
                "other"
            ],
            "x-enum-varnames": [
                "CategoryClothes",
                "CategoryFood",
                "CategoryCrafts",
                "CategoryProduce",
                "CategoryElectronics",
                "CategoryServices",
                "CategoryOther"
            ]
        },
        "entities.LoginRequest": {
            "type": "object",
            "required": [
                "password",
                "username_or_email"
            ],
            "properties": {
                "password": {
                    "description": "Password must be provided and at least 8 characters long",
                    "type": "string",
                    "minLength": 8
                },
                "username_or_email": {
                    "description": "Accepts either username or email",
                    "type": "string"
                }
            }
        },
        "entities.LoginResponse": {
            "type": "object",
            "properties": {
                "access_token": {
                    "type": "string"
                },
                "vendor_id": {
                    "type": "string"
                }
            }
        },
        "entities.Market": {
            "type": "object",
            "properties": {
                "address": {
                    "type": "string"
                },
                "close_time": {
                    "type": "string"
                },
                "created_at": {
                    "type": "string"
                },
                "deleted_at": {
                    "type": "string"
                },
                "description": {
                    "type": "string"
                },
                "id": {
                    "type": "string"
                },
                "image": {
                    "type": "string"
                },
                "latitude": {
                    "type": "string"
                },
                "layout_image": {
                    "type": "string"
                },
                "longitude": {
                    "type": "string"
                },
                "name": {
                    "type": "string"
                },
                "open_time": {
                    "type": "string"
                },
                "phone": {
                    "type": "string"
                },
                "provider": {
                    "$ref": "#/definitions/entities.MarketProvider"
                },
                "provider_id": {
                    "type": "string"
                },
                "slots": {
                    "type": "array",
                    "items": {
                        "$ref": "#/definitions/entities.Slot"
                    }
                },
                "updated_at": {
                    "type": "string"
                }
            }
        },
        "entities.MarketProvider": {
            "type": "object",
            "properties": {
                "address": {
                    "type": "string"
                },
                "email": {
                    "type": "string"
                },
                "id": {
                    "type": "string"
                },
                "markets": {
                    "type": "array",
                    "items": {
                        "$ref": "#/definitions/entities.Market"
                    }
                },
                "name": {
                    "type": "string"
                },
                "password": {
                    "type": "string"
                },
                "phone": {
                    "type": "string"
                },
                "username": {
                    "type": "string"
                }
            }
        },
...
        "entities.Slot": {
            "type": "object",
            "required": [
                "date"
            ],
            "properties": {
                "booking": {
                    "type": "array",
                    "items": {
                        "$ref": "#/definitions/entities.Booking"
                    }
                },
                "category": {
                    "$ref": "#/definitions/entities.Category"
                },
                "created_at": {
                    "type": "string"
                },
                "date": {
                    "type": "string"
                },
                "deleted_at": {
                    "type": "string"
                },
                "height": {
                    "type": "integer"
                },
                "id": {
                    "type": "string"
                },
                "market_id": {
                    "type": "string"
                },
                "name": {
                    "type": "string"
                },
                "price": {
                    "type": "number"
                },
                "status": {
                    "$ref": "#/definitions/entities.SlotStatus"
                },
                "updated_at": {
                    "type": "string"
                },
                "width": {
                    "type": "integer"
                },
                "zone": {
                    "type": "string"
                }
            }
        },
...
    },
    "securityDefinitions": {
        "BearerAuth": {
            "type": "apiKey",
            "name": "Authorization",
            "in": "header"
        }
    }
}